-

YaoBlocks.AbstractBlock — Type.

AbstractBlock

Abstract type for quantum circuit blocks.

source

YaoBlocks.AbstractContainer — Type.

AbstractContainer{BT, N, T} <: CompositeBlock{N, T}

Abstract type for container block. Container blocks are blocks contain a single block. Container block should have a

source

YaoBlocks.CacheFragment — Type.

CacheFragment{BT, K, MT}

A fragment that will be stored for each cached block (of type BT) on a cache server.

source

YaoBlocks.CachedBlock — Type.

CachedBlock{ST, BT, N, T} <: TagBlock{BT, N, T}

A label type that tags an instance of type BT. It forwards every methods of the block it contains, except mat and apply!, it will cache the matrix form whenever the program has.

source

YaoBlocks.ChainBlock — Type.

ChainBlock{N, T} <: CompositeBlock{N, T}

ChainBlock is a basic construct tool to create user defined blocks horizontically. It is a Vector like composite type.

source

YaoBlocks.CompositeBlock — Type.

CompositeBlock{N, T} <: AbstractBlock{N, T}

Abstract supertype which composite blocks will inherit from. Composite blocks are blocks composited from other AbstractBlocks, thus it is a AbstractBlock as well.

source

YaoBlocks.Concentrator — Type.

Concentrator{N, T, BT <: AbstractBlock} <: AbstractContainer{BT, N, T}

concentrates serveral lines together in the circuit, and expose it to other blocks.

source

YaoBlocks.Daggered — Type.

Daggered{N, T, BT} <: TagBlock{N, T}

Wrapper block allowing to execute the inverse of a block of quantum circuit.

source

YaoBlocks.GeneralMatrixBlock — Type.

GeneralMatrixBlock{M, N, T, MT} <: PrimitiveBlock{N, T}

General matrix gate wraps a matrix operator to quantum gates. This is the most general form of a quantum gate. M is the hilbert dimension (first dimension), N is the hilbert dimension (second dimension) of current quantum state. For most quantum gates, we have $M = N$.

source

YaoBlocks.KronBlock — Type.

KronBlock{N, T, MT<:AbstractBlock} <: CompositeBlock{N, T}

composite block that combine blocks by kronecker product.

source

YaoBlocks.Measure — Type.

Measure{N, K, OT} <: PrimitiveBlock{N, Bool}
Measure(n::Int; operator=ComputationalBasis(), locs=1:n, collapseto=nothing, remove=false)

Measure operator.

source

YaoBlocks.PhaseGate — Type.

PhiGate

Global phase gate.

source

YaoBlocks.PrimitiveBlock — Type.

PrimitiveBlock{N, T} <: AbstractBlock{N, T}

Abstract type that all primitive block will subtype from. A primitive block is a concrete block who can not be decomposed into other blocks. All composite block can be decomposed into several primitive blocks.

Note

subtype for primitive block with parameter should implement hash and == method to enable key value cache.

source

YaoBlocks.PutBlock — Type.

PutBlock <: AbstractContainer

Type for putting a block at given locations.

source

YaoBlocks.ReflectGate — Type.

ReflectGate{N, T, Tr} <: PrimitiveBlock{N, T}

Reflection operator to target state psi.

Definition

\[|ψ⟩ → 2 |s⟩⟨s| - 1\]

source

YaoBlocks.ReflectGate — Method.

ReflectGate(r::AbstractVector)

Create a ReflectGate with a quantum state vector v.

source

YaoBlocks.ReflectGate — Method.

ReflectGate(r::ArrayReg{1})

Create a ReflectGate with a quantum register r.

source

YaoBlocks.RepeatedBlock — Type.

RepeatedBlock <: AbstractContainer

Repeat the same block on given locations.

source

YaoBlocks.Roller — Type.

Roller{N, T, BT <: Tuple} <: CompositeBlock{N, T}

Roller block.

source

YaoBlocks.RotationGate — Type.

RotationGate{N, T, GT <: AbstractBlock{N, Complex{T}}} <: PrimitiveBlock{N, Complex{T}}

RotationGate, with GT both hermitian and isreflexive.

source

YaoBlocks.ShiftGate — Type.

ShiftGate <: PrimitiveBlock

Phase shift gate.

source

YaoBlocks.TagBlock — Type.

TagBlock{BT, N, T} <: AbstractContainer{BT, N, T}

TagBlock is a special kind of Container block, it forwards most of the methods but tag the block with some extra information.

source

YaoBlocks.TimeEvolution — Type.

TimeEvolution{N, TT, GT} <: PrimitiveBlock{N, ComplexF64}

TimeEvolution, where GT is block type. input matrix should be hermitian.

!!!note: TimeEvolution contructor check hermicity of the input block by default, but sometimes it can be slow. Turn off the check manually by specifying optional parameter check_hermicity = false.

source

YaoBlocks.TimeEvolution — Method.

TimeEvolution(H, dt[; tol::Real=1e-7])

Create a TimeEvolution block with Hamiltonian H and time step dt. The TimeEvolution block will use Krylove based expv to calculate time propagation.

Optional keywords are tolerance tol (default is 1e-7) TimeEvolution block can also be used for imaginary time evolution if dt is complex.

source

Base.kron — Method.

kron(n, blocks::Pair{Int, <:AbstractBlock}...)

Return a KronBlock, with total number of qubits n and pairs of blocks.

Example

Use kron to construct a KronBlock, it will put an X gate on the 1st qubit, and a Y gate on the 3rd qubit.

julia> kron(4, 1=>X, 3=>Y)
nqubits: 4, datatype: Complex{Float64}
kron
├─ 1=>X gate
└─ 3=>Y gate

source

Base.kron — Method.

kron(blocks::AbstractBlock...)
kron(n, itr)

Return a KronBlock, with total number of qubits n, and blocks should use all the locations on n wires in quantum circuits.

Example

You can use kronecker product to composite small blocks to a large blocks.

julia> kron(X, Y, Z, Z)
nqubits: 4, datatype: Complex{Float64}
kron
├─ 1=>X gate
├─ 2=>Y gate
├─ 3=>Z gate
└─ 4=>Z gate

source

Base.kron — Method.

kron(blocks...) -> f(n)
kron(itr) -> f(n)

Return a lambda, which will take the total number of qubits as input.

Example

If you don't know the number of qubit yet, or you are just too lazy, it is fine.

julia> kron(put(1=>X) for _ in 1:2)
(n -> kron(n, (n  ->  put(n, 1 => X gate)), (n  ->  put(n, 1 => X gate))))

julia> kron(X for _ in 1:2)
nqubits: 2, datatype: Complex{Float64}
kron
├─ 1=>X gate
└─ 2=>X gate

julia> kron(1=>X, 3=>Y)
(n -> kron(n, 1 => X gate, 3 => Y gate))

source

Base.repeat — Method.

repeat(x::AbstractBlock, locs)

Lazy curried version of repeat.

source

Base.repeat — Method.

repeat(n, x::AbstractBlock[, locs]) -> RepeatedBlock{n}

Create a RepeatedBlock with total number of qubits n and the block to repeat on given location or on all the locations.

source

YaoBlocks.Rx — Method.

Rx(theta)

Return a RotationGate on X axis.

source

YaoBlocks.Ry — Method.

Ry(theta)

Return a RotationGate on Y axis.

source

YaoBlocks.Rz — Method.

Rz(theta)

Return a RotationGate on Z axis.

source

YaoBlocks.apply! — Method.

apply!(register, block)

Apply a block (of quantum circuit) to a quantum register.

source

YaoBlocks.applymatrix — Method.

applymatrix(g::AbstractBlock) -> Matrix

Transform the apply! function of specific block to dense matrix.

source

YaoBlocks.cache_key — Method.

cache_key(block)

Returns the key that identify the matrix cache of this block. By default, we use the returns of parameters as its key.

source

YaoBlocks.cache_type — Method.

cache_type(::Type) -> DataType

Return the element type that a CacheFragment will use.

source

YaoBlocks.chain — Method.

chain([T=ComplexF64], n)

Return an empty ChainBlock which can be used like a list of blocks.

source

YaoBlocks.chain — Method.

chain()

Return an lambda n->chain(n).

source

YaoBlocks.chain — Method.

chain(blocks...)

Return a ChainBlock which chains a list of blocks with same nqubits and datatype. If there is lazy evaluated block in blocks, chain can infer the number of qubits and create an instance itself.

source

YaoBlocks.chcontent — Method.

chcontent(x, blk)

Create a similar block of x and change its content to blk.

source

YaoBlocks.chsubblocks — Method.

chsubblocks(composite_block, itr)

Change the sub-blocks of a CompositeBlock with given iterator itr.

source

YaoBlocks.cnot — Method.

cnot(n, ctrl_locs, location)

Return a speical ControlBlock, aka CNOT gate with number of active qubits n and locs of control qubits ctrl_locs, and location of X gate.

Example

julia> cnot(3, (2, 3), 1)
nqubits: 3, datatype: Complex{Float64}
control(2, 3)
└─ (1,) X gate

julia> cnot(2, 1)
(n -> cnot(n, 2, 1))

source

YaoBlocks.collect_blocks — Method.

collect_blocks(block_type, root)

Return a ChainBlock with all block of block_type in root.

source

YaoBlocks.concentrate — Method.

concentrate(block, locs) -> f(n)

Lazy curried version of concentrate.

source

YaoBlocks.concentrate — Method.

concentrate(n, block, locs)

Create a Concentrator block with total number of current active qubits n, which concentrates given wire location together to length(locs) active qubits, and relax the concentration afterwards.

source

YaoBlocks.content — Method.

content(x)

Returns the content of x.

source

YaoBlocks.control — Method.

control(ctrl_locs, target) -> f(n)

Return a lambda that takes the number of total active qubits as input. See also control.

Example

julia> control((2, 3), 1=>X)
(n -> control(n, (2, 3), 1 => X gate))

julia> control(2, 1=>X)
(n -> control(n, 2, 1 => X gate))

source

YaoBlocks.control — Method.

control(n, ctrl_locs, target)

Return a ControlBlock with number of active qubits n and control locs ctrl_locs, and control target in Pair.

Example

julia> control(4, (1, 2), 3=>X)
nqubits: 4, datatype: Complex{Float64}
control(1, 2)
└─ (3,) X gate

julia> control(4, 1, 3=>X)
nqubits: 4, datatype: Complex{Float64}
control(1)
└─ (3,) X gate

source

YaoBlocks.control — Method.

control(target) -> f(ctrl_locs)

Return a lambda that takes a Tuple of control qubits locs as input. See also control.

Example

julia> control(1=>X)
(ctrl_locs -> control(ctrl_locs, 1 => X gate))

julia> control((2, 3) => ConstGate.CNOT)
(ctrl_locs -> control(ctrl_locs, (2, 3) => CNOT gate))

source

YaoBlocks.control — Method.

control(ctrl_locs::Int...) -> f(target)

Return a lambda that takes a Pair of control target as input. See also control.

Example

julia> control(1, 2)
(target -> control((1, 2), target))

source

YaoBlocks.dispatch! — Method.

dispatch!(x::AbstractBlock, collection)

Dispatch parameters in collection to block tree x.

source

YaoBlocks.expect — Method.

expect(op::AbstractBlock, reg::AbstractRegister{B}) -> Vector
expect(op::AbstractBlock, dm::DensityMatrix{B}) -> Vector

expectation value of an operator.

source

YaoBlocks.getiparams — Method.

getiparams(block)

Returns the intrinsic parameters of node block, default is an empty tuple.

source

YaoBlocks.iparams_eltype — Method.

iparams_eltype(block)

Return the element type of getiparams.

source

YaoBlocks.mat — Method.

mat(blk)

Returns the matrix form of given block.

source

YaoBlocks.matblock — Method.

matblock(m::AbstractMatrix)

Create a GeneralMatrixBlock with a matrix m.

source

YaoBlocks.matblock — Method.

matblock(m::AbstractMatrix)

Create a GeneralMatrixBlock with a matrix m.

source

YaoBlocks.mathgate — Method.

mathgate(f; nbits[, bview=BitBasis.bint])

Create a MathGate with a math function f and number of bits. You can select different kinds of view which this MathGate will be applied on. Possible values are BitBasis.bint, BitBasis.bint_r, BitBasis.bfloat, BitBasis.bfloat_r.

mathgate(f; bview=BitBasis.bint) -> f(n)

Lazy curried version of mathgate.

Example

We can make a classical toffoli gate on quantum register.

julia> r = ArrayReg(bit"110")
ArrayReg{1, Complex{Float64}, Array...}
    active qubits: 3/3

julia> function toffli(b::BitStr)
           t = @inbounds b[1] ⊻ (b[3] & b[2])
           return @inbounds bit_literal(t, b[2], b[3])
       end
toffli (generic function with 1 method)

julia> g = mathgate(toffli; nbits=3)
mathgate(toffli; nbits=3, bview=bint)

julia> apply!(r, g) == ArrayReg(bit"111")
true

source

YaoBlocks.niparams — Method.

nparameters(block) -> Int

Return number of parameters in block. See also nparameters.

source

YaoBlocks.occupied_locs — Method.

occupied_locs(x)

Return an iterator of occupied locations of x.

source

YaoBlocks.parameters! — Method.

parameters!(out, block)

Append all the parameters contained in block tree with given root block to out.

source

YaoBlocks.parameters — Method.

parameters(block)

Returns all the parameters contained in block tree with given root block.

source

YaoBlocks.parameters_eltype — Method.

parameters_eltype(x)

Return the element type of parameters.

source

YaoBlocks.phase — Method.

phase(theta)

Returns a global phase gate.

source

YaoBlocks.popdispatch! — Method.

popdispatch!(block, list)

Pop the first nparameters parameters of list, then dispatch them to the block tree block. See also dispatch!.

source

YaoBlocks.popdispatch! — Method.

popdispatch!(f, block, list)

Pop the first nparameters parameters of list, map them with a function f, then dispatch them to the block tree block. See also dispatch!.

source

YaoBlocks.postwalk — Method.

postwalk(f, src::AbstractBlock)

Walk the tree and call f after the children are visited.

source

YaoBlocks.prewalk — Method.

prewalk(f, src::AbstractBlock)

Walk the tree and call f once the node is visited.

source

YaoBlocks.print_tree — Function.

print_tree(io, root, node[, depth=1, active_levels=()]; kwargs...)

Print the block tree.

Keywords

maxdepth: max tree depth to print

charset: default is ('├','└','│','─'). See also BlockTreeCharSet.

title: control whether to print the title, true or false, default is true

source

YaoBlocks.print_tree — Method.

print_tree([io=stdout], root)

Print the block tree.

source

YaoBlocks.projector — Method.

projector(x)

Return projector on 0 or projector on 1.

source

YaoBlocks.put — Method.

put(pair) -> f(n)

Lazy curried version of put.

source

YaoBlocks.put — Method.

put(total::Int, pair)

Create a PutBlock with total number of active qubits, and a pair of location and block to put on.

source

YaoBlocks.reflect — Method.

reflect(v::AbstractVector{<:Complex})

Create a ReflectGate with an quantum state vector v.

source

YaoBlocks.reflect — Method.

reflect(r::ArrayReg)

Create a ReflectGate with an ArrayReg.

source

YaoBlocks.roll — Method.

roll(n, blocks...)

Return a Roller with total number of active qubits.

source

YaoBlocks.rot — Method.

rot(U, theta)

Return a RotationGate on U axis.

source

YaoBlocks.setiparams! — Method.

setiparams!(block, itr)
setiparams!(block, params...)

Set the parameters of block.

source

YaoBlocks.setiparams! — Method.

setiparams(f, block, collection)

Set parameters of block to the value in collection mapped by f.

source

YaoBlocks.setiparams! — Method.

setiparams(f, block, symbol)

Set the parameters to a given symbol, which can be :zero, :random.

source

YaoBlocks.shift — Method.

shift(θ)

Returns a shift gate.

source

YaoBlocks.simplify — Method.

simplify(block[; rules=__default_simplification_rules__])

Simplify a block tree accroding to given rules, default to use __default_simplification_rules__.

source

YaoBlocks.subblocks — Method.

subblocks(x)

Returns an iterator of the sub-blocks of a composite block. Default is empty.

source

YaoBlocks.swap — Method.

swap([T=ComplexF64], n, loc1, loc2)

Return a n-qubit Swap gate which swap loc1 and loc2.

source

YaoBlocks.swap — Method.

swap(loc1, loc2) -> f(n)

Return a lambda that takes the total number of active qubits as input. See also swap, Swap.

source

YaoBlocks.@mathgate — Macro.

@mathgate f <nbits> <bview=bint>

Create a MathGate with a math function f and number of bits nbits, binary view bview. Unlike mathgate, f will be automatically converted to a more legible form.

Example

julia> @mathgate x->x + 0b11 nbits=4
mathgate((x -> x + 0x03); nbits=4, bview=bint)

source

YaoBlocks.HasParameters — Type.

HasParameters{X} <: SimpleTraits.Trait

Trait that block X has parameters.

source

Base.:|> — Method.

|>(register, blk)

Pipe operator for quantum circuits.

Example

julia> ArrayReg(bit"0") |> X |> Y

Warning

|> is equivalent to apply!, which means it has side effects. You need to copy original register, if you do not want to change it in-place.

source

LinearAlgebra.opnorm — Function.

opnorm(A::BlockMap, p::Real=2)

opnorm for quantum circuit blocks.

source

YaoBlocks.cunmat — Function.

cunmat(nbit::Int, cbits::NTuple{C, Int}, cvals::NTuple{C, Int}, U0::AbstractMatrix, locs::NTuple{M, Int}) where {C, M} -> AbstractMatrix

control-unitary matrix

source

YaoBlocks.decode_sign — Method.

decode_sign(ctrls...)

Decode signs into control sequence on control or inversed control.

source

YaoBlocks.getcol — Method.

getcol(csc::SDparseMatrixCSC, icol::Int) -> (View, View)

get specific col of a CSC matrix, returns a slice of (rowval, nzval)

source

YaoBlocks.print_annotation — Method.

print_annotation(io, root, node, child, k)

Print the annotation of k-th child of node, aka the k-th element of subblocks(node).

source

YaoBlocks.print_prefix — Method.

print_prefix(io, depth, charset, active_levels)

print prefix of a tree node in a single line.

source

YaoBlocks.print_title — Method.

print_title(io, block)

Print the title of given block of an AbstractBlock.

source

YaoBlocks.setcol! — Method.

setcol!(csc::SparseMatrixCSC, icol::Int, rowval::AbstractVector, nzval) -> SparseMatrixCSC

set specific col of a CSC matrix

source

YaoBlocks.u1ij! — Function.

u1ij!(target, i, j, a, b, c, d)

single u1 matrix into a target matrix.

Note

For coo, we take an additional parameter * ptr: starting position to store new data.

source

YaoBlocks.unmat — Method.

unmat(nbit::Int, U::AbstractMatrix, locs::NTuple) -> AbstractMatrix

Return the matrix representation of putting matrix at locs.

source